// source/frontend/src/models/api/generated/ApiGen_Mayan_MetadataType.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */

// LINK: @backend/apimodels/Models/Mayan/Metadata/MetadataTypeModel.cs
export interface ApiGen_Mayan_MetadataType {
  default: string | null;
  id: number;
  label: string | null;
  lookup: string | null;
  name: string | null;
  parser: string | null;
  parser_arguments: string | null;
  url: string | null;
  validation: string | null;
  validation_arguments: string | null;
}